BAU Organise files with utils folder
