+++ lambdas/delete-user-data/src/index.ts
@@ -1,11 +1,12 @@
 import { Context, SQSEvent } from "aws-lambda";
 import { deleteVCs } from "./delete-data";
-import { logger } from "./logger";
+import { logger } from "./utils/logger";
 import { readSNSMessage } from "./read-message";
-import { sendAuditEvent } from "./send-audit-event";
+import { sendAuditEvent } from "./utils/send-audit-event";
 
 export const handler = async (event: SQSEvent, context: Context): Promise<void> => {
   logger.addContext(context);
+  logger.info(JSON.stringify(event.Records), { event, recordsCount: event.Records.length });
   if (!event?.Records?.[0].body) {
     throw new TypeError("no event provided");
   }
