+++ lambdas/delete-user-data/src/utils/logger.ts
